requirements.txt
flask>=2.3.0
pdf2image>=1.16.0
pillow>=9.0.0
numpy>=1.21.0
rarfile>=4.0
py7zr>=0.20.0